cpuquiet: rqbalance: Add PM notifiers and online CPUs at resume.
This is for performance issues after system being suspended for
some time: on OSes like Android and/or userspace that will do its
synchronization, we experience sluggishness for 3 to 8 seconds,
depending on the initial load of the blocking queues.

Online CPUs and wait for the first worker complete iteration to
take over for offlining unneeded ones.

Change-Id: Ic09b38e1d541eef533f89b07ccac4b192c791da3